Change variable type from int to float

+++ bài 9 ss3 .c
@@ -1,10 +1,11 @@
 #include<stdio.h>
 #include <math.h>
 int main (){
-	int n ;
+	float n ;
 	printf("nhap n :") ;
-	scanf("%d" , &n) ;
+	scanf("%f" , &n) ;
 	float result = 1/((n - 1) * n) + 1/(n * (n + 1)) + 1/((n + 1) * (n + 2)) ;
 	printf("%.2f" , result) ;
 	return 0 ; 
 }
+
